requirements.txt
fastapi==0.116.1
starlette>=0.40.0,<1.0
uvicorn[standard]==0.30.1
pydantic>=2.9.2,<3.0.0
python-multipart>=0.0.18
jinja2==3.1.4
itsdangerous==2.2.0
requests==2.32.3
markdown2==2.4.13
google-genai
html2text